+++ resources/views/layouts/partials/footer.blade.php
@@ -1,7 +1,12 @@
 <footer>
     <p class="text-center mb-0">
-        {{ app('shaarli')->getName() }} - v{{ app('shaarli')::VERSION }} -
-        <a href="https://github.com/MarceauKa/laravel-shaarli">{{ __('Source code') }}</a>
+        {{ app('shaarli')->getName() }}
+        &mdash; v{{ app('shaarli')::VERSION }}
+        &mdash; <a href="https://github.com/MarceauKa/laravel-shaarli">{{ __('Source code') }}</a>
+        @can('restricted')
+            &mdash; <a href="{{ route('feed', 'rss') }}">{{ __('RSS Feed') }}</a>
+            &mdash; <a href="{{ route('feed', 'atom') }}">{{ __('Atom Feed') }}</a>
+        @endCan
     </p>
 </footer>