+++ routes/web.php
@@ -9,8 +9,10 @@ Route::post('login/secure/{secure}', 'SecureLoginController@check')->name('login
 Route::post('logout', 'LoginController@logout')->name('logout');
 
 Route::get('/', 'BrowseController@index')->name('home');
-Route::get('/tag/{tag}', 'BrowseController@tag')->name('tag');
+Route::get('tag/{tag}', 'BrowseController@tag')->name('tag');
 Route::get('shared/{id}/{token}', 'ShareController@view')->name('share');
+
+Route::get('feed/{type}', 'FeedController@index')->name('feed');
 
 Route::get('link/create', 'LinkController@create')->name('link.create');
 Route::get('link/edit/{id}', 'LinkController@edit')->name('link.edit');